% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyCaseData.R
\docType{data}
\name{CountyCaseData}
\alias{CountyCaseData}
\title{Number of COVID cases by county in the Upper Midwest, early 2021}
\format{
A data frame with 628 rows and 20 variables:
\describe{
\item{County}{county name}
\item{State}{state}
\item{GEOID}{County ID (for shapefiles)}
\item{37}{number of cumulative cases on Feb 6}
\item{44}{number of cumulative cases on Feb 13}
\item{51}{number of cumulative cases on Feb 20}
\item{58}{number of cumulative cases on Feb 27}
\item{65}{number of cumulative cases on Mar 6}
\item{72}{number of cumulative cases on Mar 13}
\item{79}{number of cumulative cases on Mar 20}
\item{86}{number of cumulative cases on Mar 27}
\item{93}{number of cumulative cases on Apr 3}
\item{100}{number of cumulative cases on Apr 10}
\item{107}{number of cumulative cases on Apr 17}
\item{114}{number of cumulative cases on Apr 24}
\item{121}{number of cumulative cases on May 1}
\item{128}{number of cumulative cases on May 8}
\item{135}{number of cumulative cases on May 15}
\item{142}{number of cumulative cases on May 22}
\item{149}{number of cumulative cases on May 29}
\item{MaskMandate}{day number for which mask mandate expired}
}
}
\source{
\url{https://github.com/nytimes/covid-19-data}
}
\usage{
CountyCaseData
}
\description{
A dataset containing the number of cases by county with weekly data
for Illinois, Indiana, Iowa, Michigan, Minnesota, Ohio, and Wisconsin
}
\keyword{datasets}
